+++ docs/rules/__tests__/prefer-tbody.md.spec.ts
@@ -1,0 +1,24 @@
+import HtmlValidate from "../../../src/htmlvalidate";
+
+const markup: { [key: string]: string } = {};
+markup["incorrect"] = `<table>
+	<tr><td>...</td></tr>
+</table>`;
+markup["correct"] = `<table>
+	<tbody>
+		<tr><td>...</td></tr>
+	</tbody>
+</table>`;
+
+describe("docs/rules/prefer-tbody.md", () => {
+	it("inline validation: incorrect", () => {
+		const htmlvalidate = new HtmlValidate({"rules":{"prefer-tbody":"error"}});
+		const report = htmlvalidate.validateString(markup["incorrect"]);
+		expect(report.results).toMatchSnapshot();
+	});
+	it("inline validation: correct", () => {
+		const htmlvalidate = new HtmlValidate({"rules":{"prefer-tbody":"error"}});
+		const report = htmlvalidate.validateString(markup["correct"]);
+		expect(report.results).toMatchSnapshot();
+	});
+});